backend/requirements.txt
fastapi
uvicorn[standard]
ffmpeg-python
google-generativeai
python-multipart
Pillow
requests
